fix: run gofmt; Added docstrings based on Contribution guidelines for the contributions that I've made in the past
