+++ sort/bubblesort.go
@@ -1,7 +1,7 @@
 // Implementation of basic bubble sort algorithm
 // Reference: https://en.wikipedia.org/wiki/Bubble_sort
 
-package sorts
+package sort
 
 func bubbleSort(arr []int) []int {
 	swapped := true
