+++ sort/doc.go
@@ -1,0 +1,2 @@
+// Package sort a package for demonstrating sorting algorithms in Go
+package sort
