+++ sort/selectionsort.go
@@ -1,5 +1,4 @@
-//Package sorts a package for demonstrating sorting algorithms in Go
-package sorts
+package sort
 
 func SelectionSort(arr []int) []int {
