+++ sort/shellsort.go
@@ -1,5 +1,4 @@
-//Package sorts a package for demonstrating sorting algorithms in Go
-package sorts
+package sort
 
 func ShellSort(arr []int) []int {
 	for d := int(len(arr) / 2); d > 0; d /= 2 {
