--- sorts/sorts_testcases.go
@@ -1,2 +1,0 @@
-//Package sorts a package for demonstrating sorting algorithms in Go
-package sorts
